refactor: 누락된 repository 추가 및 오타 수정 - #5
