+++ src/main/java/saphy/saphy/bookmark/domain/repository/BookmarkRepository.java
@@ -1,0 +1,8 @@
+package saphy.saphy.bookmark.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import saphy.saphy.bookmark.domain.Bookmark;
+
+public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
+
+}
