+++ src/main/java/saphy/saphy/coupon/domain/repository/CouponRepository.java
@@ -1,0 +1,8 @@
+package saphy.saphy.coupon.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import saphy.saphy.coupon.domain.Coupon;
+
+public interface CouponRepository extends JpaRepository<Coupon, Long> {
+
+}
