+++ src/main/java/saphy/saphy/delivery/domain/repository/DeliveryRepository.java
@@ -1,0 +1,8 @@
+package saphy.saphy.delivery.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import saphy.saphy.delivery.domain.Delivery;
+
+public interface DeliveryRepository extends JpaRepository<Delivery, Long> {
+
+}
