+++ src/main/java/saphy/saphy/order/domain/repository/OrderRepository.java
@@ -1,0 +1,8 @@
+package saphy.saphy.order.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import saphy.saphy.order.domain.Order;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+
+}
