+++ src/main/java/saphy/saphy/repairCenter/domain/repository/ReqairCenterRepository.java
@@ -1,0 +1,8 @@
+package saphy.saphy.repairCenter.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import saphy.saphy.repairCenter.domain.RepairCenter;
+
+public interface ReqairCenterRepository extends JpaRepository<RepairCenter, Long> {
+
+}
