+++ src/main/java/saphy/saphy/review/domain/repository/ReviewRepository.java
@@ -1,0 +1,8 @@
+package saphy.saphy.review.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import saphy.saphy.review.domain.Review;
+
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+
+}
