+++ src/main/java/saphy/saphy/user/domain/User.java
@@ -3,7 +3,6 @@ package saphy.saphy.user.domain;
 import jakarta.persistence.*;
 import lombok.*;
 import saphy.saphy.bookmark.domain.Bookmark;
-import saphy.saphy.user.domain.repository.SocialType;
 
 import java.util.List;
