+++ src/main/java/saphy/saphy/user/domain/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package saphy.saphy.user.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import saphy.saphy.user.domain.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+}
