+++ src/main/java/saphy/saphy/user/presentation/UserController.java
@@ -7,6 +7,6 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping
-public class MemberController {
+public class UserController {
 
 }
