+++ src/main/java/saphy/saphy/user/service/UserService.java
@@ -7,6 +7,6 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class MemberService {
+public class UserService {
 
 }
